requirements.txt
piexif
streamlit
tensorflow
stability-sdk
scikit-learn
pillow
pycocotools
torch
torchvision
torchaudio
fvcore
cython
opencv-python
prometheus_client